plot overwrite fixed

+++ read_data.py
@@ -1,8 +1,12 @@
 import numpy as np
 import pandas as pd
 from matplotlib import pyplot as plt
+import os
 
-tracks = pd.read_csv('data/tracks.csv')
+if os.path.isfile("data/tracks.csv"):
+    tracks = pd.read_csv('data/tracks.csv')
+    print("Tracks dataset preview")
+    print(tracks.head())
 years = np.asarray([int(y.split("-")[0]) for y in tracks['release_date']])
 tracks['year'] = years
 
@@ -12,13 +16,23 @@ tracks['decade'] = res
 
 
 
-count = dict(tracks['decade'].value_counts())
-plt.bar(x=list(count.keys()),height=list(count.values()))
-plt.savefig("./plots/decade_dist.png")
+dec_dist = dict(tracks['decade'].value_counts())
+
+if os.path.isfile('plots/decade_dist.png'):
+    dd = plt.figure()
+    plt.bar(x=list(dec_dist.keys()),height=list(dec_dist.values()))
+    plt.title("Decade Distribution")
+    plt.savefig("plots/decade_dist.png")
 
 
-tempos = tracks['tempo']
+tempos = (tracks['tempo'])
 tempos = pd.cut(tempos,8,labels=['0-50','51-75','76-100','101-125','126-150','151-175','176-200','201-250'])
 
-plt.bar(x=list(count.keys()),height=list(count.values()))
-plt.savefig('./plots/tempo_dist.png')
+tempo_dist = dict(tempos.value_counts())
+
+if os.path.isfile('plots/tempo_dist.png'):
+    plt.clf()
+    td = plt.figure()
+    plt.bar(x=list(tempo_dist.keys()),height=list(tempo_dist.values()))
+    plt.title("Tempo Distribution in bpm")
+    plt.savefig('./plots/tempo_dist.png')
